Refactor image settings from recording to new mixin
